Fuzzy search in buddy

+++ src/discordBot.py
@@ -8,6 +8,7 @@ from leaderboardBot import LeaderBoardBot
 from parseRegion import isRegion
 from dotenv import load_dotenv
 from buddies import buddyDict
+from fuzzywuzzy import process
 
 load_dotenv()
 
@@ -53,6 +54,10 @@ async def call(ctx, func, name, *args):
 
 @bot.command()
 async def buddy(ctx, *args):
+    if len(args) < 1:
+        await ctx.send("Insert a hero name after !buddy to use this command")
+        return
+
     buddyName = args[0].lower()
     try:
         await ctx.message.delete()
@@ -60,13 +65,24 @@ async def buddy(ctx, *args):
         pass
 
     if buddyName not in buddyDict.keys():
-        await ctx.send("{} is not a valid hero, try the name of the hero with no spaces or non alphabetic characters".format(buddyName))
+        buddyOptions = list(buddyDict.keys())
+        goodScores = process.extractBests(query=buddyName, choices=buddyOptions, score_cutoff=65, limit=3)
+        if len(goodScores) > 0:
+            goodScoresNames = ' or '.join(list(rate[0] for rate in goodScores))
+            await ctx.send(
+                "{} is not a valid hero, try again with {}".format(buddyName, goodScoresNames))
+        else:
+            await ctx.send("{} is not a valid hero, try the name of the hero with no spaces or non alphabetic characters".format(buddyName))
     else:
         embed = discord.Embed(title=f'{buddyDict[buddyName][0]}\'s buddy', description=buddyDict[buddyName][1])
         await ctx.send(embed=embed)
 
 @bot.command()
 async def goldenbuddy(ctx, *args):
+    if len(args) < 1:
+        await ctx.send("Insert a hero name after !goldenbuddy to use this command")
+        return
+
     buddyName = args[0].lower()
     try:
         await ctx.message.delete()
@@ -74,7 +90,14 @@ async def goldenbuddy(ctx, *args):
         pass
 
     if buddyName not in buddyDict.keys():
-        await ctx.send("{} is not a valid hero, try the name of the hero with no spaces or non alphabetic characters".format(buddyName))
+        buddyOptions = list(buddyDict.keys())
+        goodScores = process.extractBests(query=buddyName, choices=buddyOptions, score_cutoff=65, limit=3)
+        if len(goodScores) > 0:
+            goodScoresNames = ' or '.join(list(rate[0] for rate in goodScores))
+            await ctx.send(
+                "{} is not a valid hero, try again with {}".format(buddyName, goodScoresNames))
+        else:
+            await ctx.send("{} is not a valid hero, try the name of the hero with no spaces or non alphabetic characters".format(buddyName))
     else:
         embed = discord.Embed(title=f'{buddyDict[buddyName][0]}\'s golden buddy', description=buddyDict[buddyName][2])
         await ctx.send(embed=embed)
